Follow up to f7f1ec0 * Bumped HP/SP skill requirement variable up to uint32. Thanks to @CairoLee.

+++ src/map/skill.h
@@ -115,9 +115,9 @@ enum e_skill_display {
 
 /// Single skill requirement. !TODO: Cleanup the variable types
 struct skill_condition {
-	uint16 hp;								 ///< HP cost
-	uint16 mhp;								 ///< Max HP to trigger
-	uint16 sp;								 /// SP cost
+	uint32 hp;								 ///< HP cost
+	uint32 mhp;								 ///< Max HP to trigger
+	uint32 sp;								 /// SP cost
 	int16 hp_rate;							 /// HP cost (%)
 	int16 sp_rate;							 /// SP cost (%)
 	uint32 zeny;							 /// Zeny cost
